requirements.txt
aiogram==3.7.0
fastapi==0.111.0
uvicorn==0.30.1
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.5
apscheduler==3.10.4
